Added support for jRuby and improved server.
Fork is not supported in jRuby, and it was a bit of overkill since I am
just doing a fork+exec. I switched to using spawn which greatly simplifies the
code. I also changed the kill to use a TERM signal instead of a KILL. This will
ultimately make the yard server healthier.

Fixes #11